Avoid signed/unsigned warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@60331 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/sra/readers/sra/graphread.cpp
@@ -328,7 +328,7 @@ CVDBGraphSeqIterator::x_MakeGraph(const string& annot_name,
     if ( scale != 1 ) {
         graph->SetA(1./scale);
     }
-    if ( numval > kMax_Int ) {
+    if ( numval > size_t(kMax_Int) ) {
         NCBI_THROW(CSraException, eDataError,
                    "CVDBGraphSeqIterator::x_MakeGraph: graph too big");
     }
@@ -441,7 +441,7 @@ CVDBGraphSeqIterator::x_MakeTable(const string& annot_name,
         col_step->SetHeader().SetField_name("value_step");
         col_step->SetDefault().SetReal(1./scale);
     }
-    if ( arr_pos.size() > kMax_Int ) {
+    if ( arr_pos.size() > size_t(kMax_Int) ) {
         NCBI_THROW(CSraException, eDataError,
                    "CVDBGraphSeqIterator::x_MakeTable: graph too big");
     }
